fix(atlas): cluster in a dedicated 10D UMAP, and label from core members only
Prompted by davidgasquez.com/sembleverse, a really nice semantic map of the
Semble link index. It lands on the same PCA -> UMAP -> HDBSCAN spine we use,
and reading it surfaced something we had plainly wrong.

clustering space
  sembleverse fits HDBSCAN in a dedicated 10D UMAP and keeps the 2D
  projection purely for display. We were fitting HDBSCAN on the 2D display
  coords, which turns UMAP's own artifacts (tearing, crowding) into cluster
  boundaries. Scored against PCA-50 cosine space -- which neither projection
  was fit in -- the old clustering gave silhouette +0.004 coarse / -0.067
  fine. The fine tier, i.e. what you see zoomed in, was structured worse than
  chance. Clustering in 10D gives +0.033 / +0.017.

label evidence
  Chasing that led to the ~40% of points HDBSCAN calls noise. A min_samples
  sweep (1->25) shows the noise rate never leaves 36-44%, so it is real
  structure, not a tuning artifact; core points carry median membership
  probability 0.97-0.99. But silhouette on core members alone is +0.10..+0.16
  versus +0.01..+0.04 after snapping -- and the snap, which exists so every
  point gets a color, was also feeding cluster labels. Labels now come from
  core members only, with a fallback to full membership for clusters whose
  core titles are all empty. Changes 61% of coarse and 70% of fine labels.

also
  clusterFineArr was a Uint8Array while fine cluster ids run into the hundreds
  (452 live), so ids above 255 wrapped and the connection-line renderer drew
  lines between unrelated documents.

tried, not adopted
  sembleverse's other two ideas did not survive measurement at our scale,
  which is a statement about our corpus rather than about them:
  - cluster_selection_method="leaf" lost to EOM in all four matched
    comparisons and pushed noise toward 50%
  - min_cluster_size = max(100, n/100) resolves to 533 at ~53k docs and
    collapses the map to 25 clusters, breaking semantic zoom

+++ site/atlas.js
@@ -2038,7 +2038,10 @@
           uriToIndex.set(d.points[i].uri, i);
         }
         // build cluster metadata: fine cluster array, dominant platform, spatial radius
-        clusterFineArr = new Uint8Array(n);
+        // Uint16, not Uint8: fine cluster ids run into the hundreds, and
+        // truncation aliased distant clusters together (id 300 → 44), which
+        // drew connection lines between unrelated documents.
+        clusterFineArr = new Uint16Array(n);
         var coarsePlatCounts = {};
         var finePlatCounts = {};
         for (var i = 0; i < n; i++) {
@@ -2108,8 +2111,7 @@
         // Gather each fine cluster's member coords, trim far-flung outliers so
         // the cloud hugs the dense core, then stride-sample down to a small cap.
         // The render pass stamps a soft halo sprite at each of these to build an
-        // organic translucent nebula. Full int cluster ids (not the Uint8 array,
-        // which truncates past 255 fine clusters).
+        // organic translucent nebula.
         var finePts = {};
         for (var i = 0; i < n; i++) {
           var cf = d.points[i].clusterFine;
